+++ packages/atomic/src/components/common/facets/numeric-facet/values-container.tsx
@@ -1,6 +1,6 @@
 import {FunctionalComponent, h} from '@stencil/core';
 import {i18n} from 'i18next';
-import {FacetValuesGroup} from '../facet-values-group/facet-values-group';
+import {FacetValuesGroup} from '../facet-values-group/stencil-facet-values-group';
 
 interface NumericFacetValuesContainerProps {
   i18n: i18n;