+++ packages/headless/src/features/dictionary-field-context/dictionary-field-context-state.ts
@@ -1,7 +1,7 @@
 export type DictionaryFieldContextPayload = Record<string, string>;
 export type DictionaryFieldContextState = {
   /**
-   * Holds the [dictionary field context](https://docs.coveo.com/en/2036/index-content/about-fields#dictionary-fields) information.
+   * Holds the [dictionary field context](https://docs.coveo.com/en/2036/) information.
    */
   contextValues: DictionaryFieldContextPayload;
 };